requirements.txt
# =============================================================================
# Requirements for the PhysDeepSIF pipeline
# =============================================================================
# See docs/02_TECHNICAL_SPECIFICATIONS.md Section 2.1 for version details.
# Install with: pip install -r requirements.txt
# =============================================================================

# Core scientific computing
numpy>=1.24
scipy>=1.11
h5py>=3.9

# Neural simulation
tvb-library>=2.8

# EEG forward modeling
mne>=1.6

# Deep learning
torch>=2.1

# Optimization
cmaes>=0.10

# Parallelization
joblib>=1.3

# Configuration
pyyaml>=6.0

# Visualization (for validation and heatmaps)
matplotlib>=3.7
nilearn>=0.10

# Machine learning utilities
scikit-learn>=1.3

# Logging and utilities
tqdm>=4.65
